+++ Project/Client/Inspector_init.cpp
@@ -94,7 +94,7 @@ void Inspector::CreateAssetUI()
 
 void Inspector::ResizeScriptUI(UINT _Size)
 {
-	int AddSize = _Size - m_vecScriptUI.size();
+	int AddSize = (int)_Size - (int)m_vecScriptUI.size();
 
 	for (int i = 0; i < AddSize; ++i)
 	{
